Facultywork

+++ SmartStudent/src/main/java/com/manage/manager/dao/Manager_BacklogDAO.java
@@ -10,9 +10,9 @@ public interface Manager_BacklogDAO {
 
 	void updateBacklog(Backlog backlog);
 
-	void deleteBacklog(Integer studentId, String subjectCode, Integer semester);
+	void deleteBacklog(Long studentId, String subjectCode, Integer semester);
 
-	Backlog getBacklogById(Integer studentId, String subjectCode, Integer semester);
+	Backlog getBacklogById(Long studentId, String subjectCode, Integer semester);
 
 	List<Backlog> getAllBacklogs();
